+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route} from 'react-router-dom';
 import './styles/App.scss';
 import { PrivateRoute } from './_components/PrivateRoute';
-import Login from './LoginPage/Login';
+import Login from './_components/LoginPage/Login';
 import Layout from './views/Layout';
