+++ src/assets/images/defect.png
